exdata/requirements.txt
kafka-python-ng==2.2.2
pandas==2.2.3
six==1.16.0
yfinance
websockets
aiokafka

